chore(frontend): delete client.ts — workspace + csrf + debug now in dedicated v2 modules
